backend/requirements.txt
flask
pytest
pytest-mock
python-dotenv
google-genai
requests
openai
pandas
openpyxl
sentence-transformers
scikit-learn
werkzeug